Change f/o structure in translator to fix KT-48945 in FIR
The purpose of this commit is to obey the following IR invariants
in FIR2IR translator:
- first, number of fake override' overridden symbols should match
number of supertypes. This is not the case in FIR where it's allowed
to drop less concrete overridden symbols in this case
- second, fake override signature in IR should be based on the signature
of overridden symbol from the first supertype. This is also not the case
in FIR where we choose most concrete signature when we have a choice.